mypy: fix remaining reporting/charts and feature_engineering errors (20 -> 0)
fuzzy_features.py, recency_density.py, hurst.py, spectral.py: float()/
np.asarray() wraps on njit-kernel returns per the no-any-return
pattern; _spec_pow's declared return annotation corrected from
np.ndarray to Tuple[np.ndarray, np.ndarray] (a real annotation bug --
the function has always returned a 2-tuple).

timeseries.py: guard splitting_ratios_names writes/reads on
create_features_names AND is-not-None (the Optional[list] contract
was only half-checked).

calibration_heatmap_2d.py: np.asarray() on feat_x/feat_y before
compute_calibration_heatmap_2d (accepts Sequence at the outer API,
requires ndarray internally).

risk_coverage.py: Tuple[Tuple[float,float,str,str,str], ...]
annotation on the point_markers accumulator (was narrowed to the
empty-tuple type by its first assignment).

diagnostics_dispatch.py: list()-wrap on the ndarray timestamps slice
passed to worst_k_table's Sequence-typed param, and renamed the
second try block's `res` (ErrorBiasResult) to eb_res to stop
colliding with the earlier WeakSegmentResult-typed `res`.

regression.py: narrowed trend_line's Optional[str] to
Optional[Literal["theil-sen","huber"]] across three call sites (the
only two values the panel specs' render path accepts).

split_comparison.py: list()-wrap on two dict.keys() views passed to
_order_splits's Sequence[str] param.

+++ src/mlframe/feature_engineering/hurst.py
@@ -48,7 +48,7 @@ def compute_hurst_rs(arr: np.ndarray) -> float:  # pragma: no cover
     S = np.sqrt(var)
     if R <= _ZERO_EPS or S <= _ZERO_EPS:
         return np.nan
-    return R / S
+    return float(R / S)
 
 
 @njit(fastmath=_FASTMATH, cache=True)
@@ -191,7 +191,7 @@ def _hurst_rs_single(x: np.ndarray) -> float:
     sd = x.std()
     if sd <= _ZERO_EPS or rng_v <= _ZERO_EPS:
         return np.nan
-    return np.log(rng_v / sd) / np.log(n)
+    return float(np.log(rng_v / sd) / np.log(n))
 
 
 @njit(cache=True, fastmath=True, parallel=True)
